Send chunked responses by default
Using chunked requests, it is possible for a client/server to send the message body to the message receiver in chunks. Also, per RFC 9112, Section 7.1-4, "A recipient MUST be able to parse and decode the chunked transfer coding.", which means that every client/server that conforms to RFC 9112 MUST be able to handle chunked messages.

Send Transfer-Encoding header with value "chunked" by default

Update examples accordingly